Improved type definitions for react-widgets
